third_party/blink/web_tests/platform/mac-mac-arm11.0/external/wpt/html/canvas/offscreen/text/2d.text.draw.align.start.rtl.worker-expected.txt

This is a testharness.js-based test.
FAIL textAlign start with rtl is the right edge assert_approx_equals: Red channel of the pixel at (5, 5) expected 0 +/- 2 but got 255
Harness: the test ran to completion.
